backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
python-multipart>=0.0.9
pillow>=10.0
numpy>=1.24
paddleocr>=2.7
paddlepaddle>=2.5
pytest>=7.4
